Updated nxnk.py
Fix Graph.degree to present an interface more similar to networkx,
with the "node" argument being a keyword argument nbunch=None.

Unlike networkx, if nbunch is an iterable nothing will happen.

degree is one of the few functions that fundamentally difers in behavior
from networkx, in that it refuses to handle arguments of different type
(a single node, or an iterable of nodes).