Updated nxnk.py
New classes NXGraph and NXDiGraph, which extend Graph and DiGraph respectively,
with methods that behave like those of the homonimous networkx classes.

For example, nodes() returns a list instead of a generator.

These classes enable using nxnk graphs in networkx library functions without modification,
as drop-in replacements for networkx's Graph and DiGraph classes.

Graph and DiGraph acquired a new method, nx_adapter, which returns
an editable view of themselves (not a copy) as NXGraph and NXDiGraph,
respectively.